The edge device component of the AtomPusher AWS IoT system
This piece is intended to run on the edge device, the sensor that is
out in the world. In this case, it’s a simple javascript program that
grabs an atom feed, applies a custom tag that you specify, and then
sends it along on the ‘feeds’ topic into the system. From there it gets
processed by the Lambda component and ends up in the Collector.